+++ assets/js/forms.js
@@ -1,57 +1,11 @@
-(function () { var require = undefined; var define = undefined; (function(){function r(e,n,t){function o(i,f){if(!n[i]){if(!e[i]){var c="function"==typeof require&&require;if(!f&&c)return c(i,!0);if(u)return u(i,!0);var a=new Error("Cannot find module '"+i+"'");throw a.code="MODULE_NOT_FOUND",a}var p=n[i]={exports:{}};e[i][0].call(p.exports,function(r){var n=e[i][1][r];return o(n||r)},p,p.exports,r,e,n,t)}return n[i].exports}for(var u="function"==typeof require&&require,i=0;i<t.length;i++)o(t[i]);return o}return r})()({1:[function(require,module,exports){
+(function(){function r(e,n,t){function o(i,f){if(!n[i]){if(!e[i]){var c="function"==typeof require&&require;if(!f&&c)return c(i,!0);if(u)return u(i,!0);var a=new Error("Cannot find module '"+i+"'");throw a.code="MODULE_NOT_FOUND",a}var p=n[i]={exports:{}};e[i][0].call(p.exports,function(r){var n=e[i][1][r];return o(n||r)},p,p.exports,r,e,n,t)}return n[i].exports}for(var u="function"==typeof require&&require,i=0;i<t.length;i++)o(t[i]);return o}return r})()({1:[function(require,module,exports){
 'use strict';
 
-var _scrollToElement = _interopRequireDefault(require("./misc/scroll-to-element.js"));
-
 require("./forms/conditional-elements.js");
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
 
 var mc4wp = window.mc4wp || {};
 
 var forms = require('./forms/forms.js');
-
-var config = window.mc4wp_forms_config || {};
-
-function handleFormRequest(form, eventName, errors, data) {
-  var timeStart = Date.now();
-  var pageHeight = document.body.clientHeight; // re-populate form if an error occurred
-
-  if (errors) {
-    form.setData(data);
-  } // scroll to form
-
-
-  if (window.scrollY <= 10 && config.auto_scroll) {
-    (0, _scrollToElement["default"])(form.element);
-  } // trigger events on window.load so all other scripts have loaded
-
-
-  window.addEventListener('load', function () {
-    trigger('submitted', [form]);
-
-    if (errors) {
-      trigger('error', [form, errors]);
-    } else {
-      // form was successfully submitted
-      trigger('success', [form, data]); // subscribed / unsubscribed
-
-      trigger(eventName, [form, data]); // for BC: always trigger "subscribed" event when firing "updated_subscriber" event
-
-      if (eventName === 'updated_subscriber') {
-        trigger('subscribed', [form, data, true]);
-      }
-    } // scroll to form again if page height changed since last scroll, eg because of slow loading images
-    // (only if load didn't take too long to prevent overtaking user scroll)
-
-
-    var timeElapsed = Date.now() - timeStart;
-
-    if (config.auto_scroll && timeElapsed > 1000 && timeElapsed < 2000 && document.body.clientHeight !== pageHeight) {
-      (0, _scrollToElement["default"])(form.element);
-    }
-  });
-}
 
 function trigger(event, args) {
   forms.trigger(args[0].id + "." + event, args);
@@ -110,19 +64,11 @@ if (mc4wp.listeners) {
 } // expose forms object
 
 
-mc4wp.forms = forms; // handle submitted form
-
-if (config.submitted_form) {
-  var formConfig = config.submitted_form,
-      element = document.getElementById(formConfig.element_id),
-      form = forms.getByElement(element);
-  handleFormRequest(form, formConfig.event, formConfig.errors, formConfig.data);
-} // expose mc4wp object globally
-
+mc4wp.forms = forms; // expose mc4wp object globally
 
 window.mc4wp = mc4wp;
 
-},{"./forms/conditional-elements.js":2,"./forms/forms.js":4,"./misc/scroll-to-element.js":5}],2:[function(require,module,exports){
+},{"./forms/conditional-elements.js":2,"./forms/forms.js":4}],2:[function(require,module,exports){
 'use strict';
 
 function getFieldValues(form, fieldName) {
@@ -271,7 +217,7 @@ Form.prototype.reset = function () {
 
 module.exports = Form;
 
-},{"form-serialize":6,"populate.js":7}],4:[function(require,module,exports){
+},{"form-serialize":5,"populate.js":6}],4:[function(require,module,exports){
 'use strict';
 
 var Form = require('./form.js');
@@ -356,28 +302,6 @@ module.exports = {
 };
 
 },{"./form.js":3}],5:[function(require,module,exports){
-'use strict';
-
-function scrollTo(element) {
-  var x = window.pageXOffset || document.documentElement.scrollLeft;
-  var y = calculateScrollOffset(element);
-  window.scrollTo(x, y);
-}
-
-function calculateScrollOffset(elem) {
-  var body = document.body,
-      html = document.documentElement;
-  var elemRect = elem.getBoundingClientRect();
-  var clientHeight = html.clientHeight;
-  var documentHeight = Math.max(body.scrollHeight, body.offsetHeight, html.clientHeight, html.scrollHeight, html.offsetHeight);
-  var scrollPosition = elemRect.bottom - clientHeight / 2 - elemRect.height / 2;
-  var maxScrollPosition = documentHeight - clientHeight;
-  return Math.min(scrollPosition + window.pageYOffset, maxScrollPosition);
-}
-
-module.exports = scrollTo;
-
-},{}],6:[function(require,module,exports){
 // get successful control from form and assemble into object
 // http://www.w3.org/TR/html401/interact/forms.html#h-17.13.2
 
@@ -639,7 +563,7 @@ function str_serialize(result, key, value) {
 
 module.exports = serialize;
 
-},{}],7:[function(require,module,exports){
+},{}],6:[function(require,module,exports){
 /*! populate.js v1.0.2 by @dannyvankooten | MIT license */
 ;(function(root) {
 
@@ -736,4 +660,3 @@ module.exports = serialize;
 }(this));
 
 },{}]},{},[1]);
- })();