+++ includes/forms/views/js/general-form-enhancements.js
@@ -1,7 +1,0 @@
-function addEventListener(element,event,handler) {
-	if (element.addEventListener) {
-		element.addEventListener( event,handler, false );
-	} else if (element.attachEvent) {
-		element.attachEvent( 'on' + event,handler );
-	}
-}